First cut headless impl

+++ cli/termcast.py
@@ -36,7 +36,7 @@ class Host(object):
         return 'http%s://%s/' % ('s' if self._ssl else '', self._domain)
 
 
-def communicate(host, session, fifo, output, tmux_socket):
+def communicate(host, session, fifo, output, tmux_socket=None):
     """Handle all the bidirectional traffic"""
 
     #TODO: This should be in a config file.
@@ -60,13 +60,13 @@ def communicate(host, session, fifo, output, tmux_socket):
         with connection_lock:
             while container['connection'] is None or not container['connection'].connected:
                 try:
-                    logging.warn('Attempting websocket connection')
+                    logging.info('Attempting websocket connection')
                     out.write('Connecting...\n')
                     container['connection'] = create_connection(url)
                     container['connection'].send(json.dumps({'type': 'registerPublisher',
                                                              'token': session['token'],
                                                              'body': ''}))
-                    logging.warn('Attempt succeeded')
+                    logging.info('Attempt succeeded')
                     out.write(template % (session['id'], container['viewers']))
                 except Exception:
                     logging.error('Attempt to create websocket connection failed')
@@ -112,7 +112,7 @@ def communicate(host, session, fifo, output, tmux_socket):
                 if received['type'] == 'viewcount':
                     container['viewers'] = received['body']
                     out.write(template % (session['id'], container['viewers']))
-                if received['type'] == 'requestSync':
+                if received['type'] == 'requestSync' and tmux_socket != None:
                     sync(target=received['requester'])
 
             except Exception:
@@ -177,13 +177,14 @@ def do_the_needful():
     parser.add_argument('--session', default=None)
     parser.add_argument('--logfile', default=None)
     parser.add_argument('--host', default='https://termcast.me')
+    parser.add_argument('--command', type=str, default=None)
     args = parser.parse_args()
 
     unique_id = uuid.uuid4()
 
     # Configure logging
     if args.logfile is not None:
-        logging.basicConfig(level=logging.INFO,
+        logging.basicConfig(level=logging.ERROR,
                             filename=args.logfile,
                             format='[%(asctime)s] {%(pathname)s:%(lineno)d} %(levelname)s - %(message)s',
                             datefmt='%H:%M:%S')
@@ -241,30 +242,54 @@ def do_the_needful():
             traceback.print_exc()
             exit(1)
 
-    comms_thread = Thread(target=communicate, args=(host, session, fifo, output, tmux_socket))
-    comms_thread.daemon = True
-    comms_thread.start()
-
     if platform.system() == 'Darwin':
         flush = '-F'
     else:
         flush = '-f'
 
-    subprocess.call(['tmux', '-S', tmux_socket, '-2', '-f', tmux_config,
-                     'new', 'script', '-q', '-t0', flush, fifo])
-    time.sleep(0.1) # Give tmux a chance to start - should really wait for socket file to exist.
+    if args.command != None:
+        f = os.fork()
+        if f == 0:
+            devnull = open('/dev/null', 'w')
+            proc = subprocess.Popen(['script', '-q', '-t0', flush, fifo] + args.command.split(' '),
+                                   stdout=devnull)
+            print('%s/%s' % (args.host, session['id']), proc.pid)
 
-    # Tidy up after ourselves.
-    for mess in [fifo, tmux_config, output, tmux_socket]:
-        os.remove(mess)
+            comms_thread = Thread(target=communicate, args=(host, session, fifo, output))
+            comms_thread.daemon = True
+            comms_thread.start()
 
-    print("You have disconnected from your broadcast session.")
-    print("To reconnect, run:")
-    print("")
-    print("termcast --session %s --token %s --width %s --height %s" % (session['id'], 
-                                                                       session['token'],
-                                                                       session['width'],
-                                                                       session['height']))
+            proc.communicate()
+
+            # Tidy up after ourselves.
+            for mess in [fifo, tmux_config, output]:
+                os.remove(mess)
+
+        else:
+            exit(0)
+
+    else:
+        proc = subprocess.Popen(['tmux', '-S', tmux_socket, '-2', '-f', tmux_config,
+                                 'new', 'script', '-q', '-t0', flush, fifo])
+
+        comms_thread = Thread(target=communicate, args=(host, session, fifo, output, tmux_socket))
+        comms_thread.daemon = True
+        comms_thread.start()
+
+        proc.communicate()
+
+        print("You have disconnected from your broadcast session.")
+        print("To reconnect, run:")
+        print("")
+        print("termcast --session %s --token %s --width %s --height %s" % (session['id'], 
+                                                                           session['token'],
+                                                                           session['width'],
+                                                                           session['height']))
+
+        # Tidy up after ourselves.
+        for mess in [fifo, tmux_config, output, tmux_socket]:
+            os.remove(mess)
+
 
 if __name__ == "__main__":
     do_the_needful()
